Added Tetrimino Builder

+++ src/main/java/domergue/bastide/jTetris/components/tetriminos/Tetrimino.java
@@ -2,11 +2,11 @@ package main.java.domergue.bastide.jTetris.components.tetriminos;
 
 public class Tetrimino {
 	
-	private int tetriminoId;
+	private long tetriminoId;
 	private Shape shape;
 	
 	
-	public Tetrimino(int tetriminoId, Shape shape){
+	public Tetrimino(long tetriminoId, Shape shape){
 		this.tetriminoId = tetriminoId;
 		this.shape = shape;
 	}
@@ -17,10 +17,10 @@ public class Tetrimino {
 	public void setShape(Shape shape) {
 		this.shape = shape;
 	}
-	public int getTetriminoId() {
+	public long getTetriminoId() {
 		return tetriminoId;
 	}
-	public void setTetriminoId(int tetriminoId) {
+	public void setTetriminoId(long tetriminoId) {
 		this.tetriminoId = tetriminoId;
 	}
